% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_matrix.R
\name{choose_matrix}
\alias{choose_matrix}
\title{Choose Covariance Matrix Structure}
\usage{
choose_matrix(sig2, rho = 0.7, p, blk_sz = 20, structure)
}
\arguments{
\item{sig2}{Scalar variance of each covariate (assumed constant across covariates).}

\item{rho}{Correlation for AR-1 structure (and block-diagonal structure with AR-1
blocks).}

\item{p}{Covariance matrix dimension.}

\item{blk_sz}{Size of blocks for block-diagonal structure.}

\item{structure}{Covariance matrix structure (diag, block, or AR1).}
}
\value{
Covariance matrix
}
\description{
Returns a covariance matrix of the specified dimension and structure.
}
