% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_gen_lmm_bsl.R
\name{data_gen_lmm_bsl}
\alias{data_gen_lmm_bsl}
\title{Generate Data from Linear Mixed-Effects Model with Constant Covariates.}
\usage{
data_gen_lmm_bsl(
  n,
  m,
  p1,
  p2,
  q,
  k,
  mu,
  beta_t,
  beta1,
  beta2,
  mu_x,
  Sigma_x,
  Omega_0,
  Sigma_d1,
  z,
  Sigma_e,
  theta
)
}
\arguments{
\item{n}{Number of clusters (e.g. individuals, clinics, etc.).}

\item{m}{Number of repeated measures per cluster (assumed equal across clusters).}

\item{p1}{Number of error-prone covariates.}

\item{p2}{Number of error-free covariates.}

\item{q}{Number of variance components.}

\item{k}{Scalar constant or n-dimensional vector of the number of replicates
for each cluster. If a scalar is provided, the number of replicates is assumed
to be constant across clusters.}

\item{mu}{Population mean for the LMM}

\item{beta_t}{Scalar coefficient for the time effect.}

\item{beta1}{p1-dimensional coefficient vector for error-prone covariates.}

\item{beta2}{p2-dimensional coefficient vector for error-free covariates.}

\item{mu_x}{(p1+p2)-dimensional vector for the mean of the covariate distribution.}

\item{Sigma_x}{(p1+p2)x(p1+p2)-dimensional covariance matrix for covariates.}

\item{Omega_0}{q x q dimensional covariance matrix for random effects.}

\item{Sigma_d1}{p1 x p1 dimensional covariance matrix for measurement errors.}

\item{z}{Index vector of covariates to include in random effects matrix. -1
indicates a random intercept, and 0 indicates a random slope associated with
time. All other positive integers correspond to columns of the covariate matrix
X.}

\item{Sigma_e}{m x m dimensional covariance matrix for residual errors.}

\item{theta}{d-dimensional vector of variance components for the random effects
covariance matrix.}
}
\value{
a list with generated data and metadata. Importantly, Xs is the generated
error-prone data of dimension sum(k) x (p1+p2). The first k_1 rows correspond to
the k_1 replicates for cluster 1, the next k_2 to the k_2 replicates for cluster
2, and so on.
}
\description{
Generates covariates, as well as error-prone versions, and produces a response
vector based on a specified Linear Mixed-Effects Model. The generated data
and metadata are returned in a list. Covariates are assumed to be constant.
}
