% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtheta.R
\name{devtheta}
\alias{devtheta}
\title{Deviance function (twice negative log-likelihood) parametrized by variance
components.}
\usage{
devtheta(
  theta,
  y,
  X,
  Z,
  L,
  Lambda,
  sig2e,
  uk,
  beta,
  n,
  N,
  p,
  q,
  R_XtR_X = NULL,
  REML = TRUE
)
}
\description{
Deviance function (twice negative log-likelihood) parametrized by variance
components.
}
