% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iclmmLasso.R
\name{ic.lmm_lasso}
\alias{ic.lmm_lasso}
\title{Coordinate-descent LMM Lasso Algorithm with AIC/BIC-based Tuning Parameter Selection}
\usage{
ic.lmm_lasso(
  formula,
  data,
  p,
  q,
  d,
  n,
  N,
  theta_0 = NULL,
  beta_0 = NULL,
  sig2e_0 = NULL,
  lambda.max = NULL,
  nlambda = 20,
  lambda.min.ratio = 0.001,
  penalty.factor = NULL,
  REML = TRUE,
  BIC = TRUE,
  eta = 1e-04,
  eta_gs = 0.005,
  maxits = 50000,
  standardize = TRUE,
  progress = FALSE,
  getInit = FALSE,
  ...
)
}
\arguments{
\item{formula}{LMM formula to fit}

\item{data}{data frame which includes the variables from the formula}

\item{p}{dimension of covariate matrix X (also, number of fixed effects)}

\item{q}{number of random effects}

\item{d}{variance components vector dimension}

\item{n}{number of clusters}

\item{N}{total number of response observations, including repeated measures}

\item{theta_0}{initial estimate for variance components vector (optional)}

\item{beta_0}{initial estimate for fixed-effects coefficient vector (optional)}

\item{sig2e_0}{initial estimate for residual error variance (optional)}

\item{lambda.max}{maximum value for penalty parameter over which to optimize.
If not provided, one will be computed using a data-driven approach.}

\item{nlambda}{Number of penalty parameters to search through.}

\item{lambda.min.ratio}{Ratio which will determine minimum lambda (lambda min =
lambda.max*lambda.min.ratio). Default is 0.001.}

\item{penalty.factor}{Optional p-dimensional vector of penalty weights, so that
different predictors can have different penalties. Default is equal weighting, and
weights are normalized to sum to 1.}

\item{REML}{specify whether the ML or REML estimator will be used (default TRUE)}

\item{BIC}{specify whether BIC or AIC will be returned for the fitted model.}

\item{eta}{Threshold for convergence of outer loop (which solves the mixed model equations)}

\item{eta_gs}{Threshold for convergence of inner loop (which performs coordinate descent)}

\item{maxits}{Maximum iterations before exit}

\item{standardize}{specify whether to standardize predictors (i.e. mean-center and
scale to have unit variance) for the fitting procedure. Final coefficient estimates
will be rescaled to the original predictor scales.}

\item{progress}{0, 1, or 2. Larger integers will result in printing more information
on the progress of the algorithm.}

\item{getInit}{specify whether you would like just the data driven choice of lambda
and the computed initial values for theta, beta, and sig2e to be returned. If so,
the algorithm will not be run to completion; only the initial values will be
computed and returned. Default is FALSE.}

\item{...}{extra parameters for lmer function}
}
\value{
list with AIC/BIC value, variance component estimate, random effects
covariance matrix estimate, intercept, fixed-effects, and random-effects
estimates, residual error variance estimate, optimal penalty parameter, and
matrix of indicators S_hat. S_hat is an nlambda x p matrix with 1's where the
model produced a nonzero coefficient estimate, and 0's elsewhere.
}
\description{
Coordinate-descent based algorithm to fit the LMM Lasso with AIC/BIC-based tuning parameter selection.
}
