% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapping}
\alias{mapping}
\title{Mapping of variance components to random effects covariance matrix}
\usage{
mapping(theta, n, q, init = TRUE)
}
\description{
Mapping of variance components to random effects covariance matrix
}
